add function to update multiple param values at once
